requirements.txt
# ============================================================
# Smart Waste Segregation System - Requirements
# ============================================================

# Core Deep Learning
tensorflow>=2.12.0
keras>=2.12.0

# Data Science & Visualization
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0
scikit-learn>=1.2.0

# Image Processing
Pillow>=9.4.0
opencv-python>=4.7.0

# Web Application
streamlit>=1.22.0

# Voice Output
pyttsx3>=2.90

# Utilities
tqdm>=4.64.0
requests>=2.28.0
gdown>=4.7.0       # for Google Drive dataset downloads

# Optional: GPU support
# tensorflow-gpu>=2.12.0
